requirements.txt
uuid
typing
numpy
fastapi
pydantic
pypdf
python-docx
sentence_transformers
openai
uvicorn